model/requirements.txt
# PyTorch with CUDA — install the correct wheel for your CUDA version:
# CUDA 12.1:  pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# CUDA 11.8:  pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# CPU only:   pip install torch torchvision
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
onnx>=1.15.0
